// src/app/services/models/reclamation-response.ts
/* tslint:disable */
/* eslint-disable */
export interface ReclamationResponse {
  contenu?: string;
  etat?: 'En cours de traitement' | 'Trait\xE9';
  firstname?: string;
  gmail?: string;
  idReclamation?: number;
  lastname?: string;
  reponse?: string;
  type?: 'ProblemeEnchere' | 'ProblemePaiement' | 'Problemecompteutilisateur' | 'Problemetechnique';
}